Show articles in db using ListView and CursorAdapter
Made a custom CursorAdapter that takes value from a cursor and inserts it into
custom made list item layout (article_list_item).

Added a ListView to MainActivity which uses this CursorAdapter.

When a readableDatabase is returned from the AsyncTask, the cursorAdapter is set
to the listView. This was checked by using sqlite3 utility present in adb.

Also, when the application is close, the cursor is closed.

Signed-off-by: Neha Bagari <neha.bagari.eee11@iitbhu.ac.in>