Save the articles in database
The articles which were received from the api calls on pressing refresh button
are now saved in the database.
The latest timestamp is stored in SharedPreferences to sent in further requests.

But there is some inconsitency in timestamp logic. What should happen in case
there is an error in inserting the article to the database? Which timestamp
should be saved in the preferences?

I have just included a barely working model. This needs to be overhauled or
would lead to inconsistencies in the database.

Made changes to DbHelper. Key class is made unique.

Signed-off-by: Neha Bagari <neha.bagari.eee11@iitbhu.ac.in>