Hidden field containing article key
A clever way to extract key from List Item.
A hidden TextView is created in which the key of the article is inserted via the
CursorAdapter.
This is then extracted in OnItemClickListener by iterating over child views and
comparing their ids. The TextView with the id tv_list_key (id of hiddent
TextView) is chosen and its text is extracted. This would be passed on to the
new activity.

Signed-off-by: Neha Bagari <neha.bagari.eee11@iitbhu.ac.in>