Fetch the news article and show it
Using the API calls, the news articles are fetched and shown to the user.
These articles are not yet cached (stored in db) nor their existence are looked
up in local db first. (To be done in next commit)

Also made some UI changes for better view of the Article.

Signed-off-by: Neha Bagari <neha.bagari.eee11@iitbhu.ac.in>